kretanje-klijent/requirements.txt
aioconsole==0.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pillow==11.1.0
requests==2.32.3
RPi.GPIO==0.7.1
urllib3==2.3.0
websockets==10.4
